fixes issue #5
To check if it was already logged in it would check if it had to approve the SSO for upay. The approval for upay lasted longer than upay's login and hence it would try and get the balance from the upay log in screen. I have changed the way it checks it is logged in and separated the SSO confirmation and upay login sections.