Redone the way it downloads new menus and animation
The notifications service and widget menu have been rewritten to avoid
blocking the main thread when there is an old menu. The file is still
read on the main thread I think and this may need to be moved to support
weaker devices. The adding and removing animation now plays after the
switch animation based on the end of the animation instead of by
delaying the start. Fixed some issues of the animation on rotate.