The twitter feed now caches images.
It checks if the image is already saved to device, if it isn't the image
is downloaded and saved to the device. The image is then read from the
device. If an image isn't used in a refresh it is removed from the
device.

Also removed a couple of printlns.